#27 refactor: move util function to libft
